Add `get_encoder(dtype, shape)` function
The encoder can no longer be selected based only on the data type because of a change in the datatype definition in the API. Now int, floats and numpy arrays are considered 'Numeric', but numbers and numpy arrays require different types of encoders and logic to validate shape and type.

To deal with the change we added a encoder selector function based on datatype and shape.